+++ src/_modules/auth/auth.service.ts
@@ -28,24 +28,6 @@ class AuthService {
     }
 
 
-    // Clean full User object, removing sessions, password, OAuth IDs and profile
-    async sanitizeUserObject(user: User, opts?:{id?:boolean, verified?:boolean, active?:boolean, comment?:boolean}){
-        const cleanUser: Partial<User> = {
-            id: opts?.id ? user.id : undefined,
-            firstname: user.firstname,
-            lastname: user.lastname,
-            roles: user.roles,
-            email: user.email,
-            emailVerified: opts?.verified ? user.emailVerified : undefined,
-            phone: user.phone,
-            profileId: user.profileId,
-            isActive: opts?.active ? user.isActive : undefined,
-            isComment: opts?.comment ? user.isComment : undefined,
-            createdAt: user.createdAt
-        };
-
-        return cleanUser;
-    }
 
     /**Validate email and password */
     validateCredentials(email:string, password:string, confirmPassword?:string){